clang-legacy.bbclass: Remove -fuse-ld=lld from SELECTED_OPTIMIZATION when using LTO.
Referring to lto.inc in openembedded-core, when using lto, the LTO variable is added to SELECTED_OPTIMIZATION.
In this case, with the clang toolchain, -fuse-ld=lld is added when using -flto,
but the -fuse-ld=lld option is sufficient as a linker flag.

If -fuse-ld=lld is present in CFLAGS, an unused-command-line-argument warning occurs.
In recipes that treat this warning as an error, the build fails.
aarch64-yoe-linux-clang: error: argument unused during compilation: '-fuse-ld=lld' [-Werror,-Wunused-command-line-argument]

See more details https://errors.yoctoproject.org/Errors/Details/886599

Signed-off-by: mark.yang <mark.yang@lge.com>